All Completed upto Arrays II

+++ DSA - C++/Algorithms/kadanesAlgorithm.cpp
@@ -1,10 +1,12 @@
 //Max Sub-Array Sum
 
 #include <iostream>
+#include <vector>
+#include <climits>
 using namespace std;
 
 int maxSubarraySum(int array[], int size){
-    int maxSum = INT16_MIN, currSum = 0;
+    int maxSum = INT_MIN, currSum = 0;
     for(int i=0; i<size; i++){
         currSum += array[i];
         maxSum = max(maxSum, currSum);
@@ -15,6 +17,23 @@ int maxSubarraySum(int array[], int size){
     }
     return maxSum;
 }
+
+
+// int maxProduct(vector<int> nums) {
+//     int currMax = nums[0], currMin = nums[0], maxProd = nums[0];
+//     for(int i=1; i<nums.size(); i++){
+//         if(nums[i]<0){
+//             swap(currMax, currMin);
+//         }
+
+//         currMax = max(nums[i], nums[i]*currMax);
+//         currMin = min(nums[i], nums[i]*currMin);
+
+//         maxProd = max(maxProd, currMax);
+//     }
+//     return maxProd;
+// }
+
 
 int main(){
     int array[] = {2,-3,6,-5,4,2};
